resolving dependencies
